refactor: split three files my changes pushed over the size cap
Engine activity + the attention Inbox contracts move beside the automation
ones; hosted-session's engine-readiness probing becomes its own module and is
re-exported; the runner specs split into when-it-fires and what-gets-recorded
over shared fixtures. No behavior change.
