+++ packages/kobe/test/daemon/automation-runner.test.ts
@@ -2,8 +2,6 @@ import { mkdtempSync, writeFileSync } from "node:fs"
 import { tmpdir } from "node:os"
 import { join } from "node:path"
 import { describe, expect, it, vi } from "vitest"
-import type { DaemonRpcClient } from "../../../kobe-daemon/src/client/rpc.ts"
-import type { DispatchRuntime } from "../../../kobe-daemon/src/daemon/automation-dispatch.ts"
 import {
   dueAutomations,
   resolveDueOccurrence,
@@ -12,101 +10,7 @@ import {
   sweepAutomations,
 } from "../../../kobe-daemon/src/daemon/automation-runner.ts"
 import { AutomationsStore } from "../../../kobe-daemon/src/daemon/automations-store.ts"
-import type { Automation, DaemonTask } from "../../../kobe-daemon/src/daemon/contracts.ts"
-
-const NOW = new Date(2026, 6, 31, 10, 0, 0).getTime() // Fri 2026-07-31 10:00 local
-const HOUR = 3_600_000
-/** A real directory: prechecks spawn with `cwd: automation.repo`. */
-const REPO = process.cwd()
-
-function automation(overrides: Partial<Automation> = {}): Automation {
-  return {
-    id: "auto-1",
-    name: "audit",
-    repo: REPO,
-    prompt: "run the audit",
-    schedule: "0 9 * * *",
-    enabled: true,
-    nextRunAt: new Date(NOW).toISOString(),
-    missedRunGraceMinutes: 60,
-    createdAt: new Date(NOW - 30 * 24 * HOUR).toISOString(),
-    updatedAt: new Date(NOW - 30 * 24 * HOUR).toISOString(),
-    ...overrides,
-  }
-}
-
-async function tempStore(now = () => NOW): Promise<AutomationsStore> {
-  const dir = mkdtempSync(join(tmpdir(), "kobe-automation-runner-"))
-  const store = new AutomationsStore(join(dir, "automations.json"), now)
-  await store.init()
-  return store
-}
-
-const link = {} as DaemonRpcClient
-
-function fakeDeps(args: {
-  store: AutomationsStore
-  createTask?: (input: unknown) => Promise<DaemonTask>
-  start?: DispatchRuntime["startTaskSessionWithPrompt"]
-  /** Tasks `resolveStandingTask` can find, keyed by id. */
-  tasks?: Record<string, DaemonTask>
-  deliver?: DispatchRuntime["deliverPromptToLiveEngineDetailed"]
-}) {
-  const created: unknown[] = []
-  const prompts: string[] = []
-  const delivered: string[] = []
-  /** One slot per routine — the store's real dedupe shape, so the test can
-   *  tell "refreshed the same episode" from "filed another one". */
-  const episodes = new Map<string, { name: string; status: string; error?: string; taskId: string | null }>()
-  const inbox = {
-    recordPromptDeferred: async () => {},
-    recordRoutineFailure: async (
-      routine: { automationId: string; name: string; status: string; error?: string },
-      taskId: string | null,
-    ) => {
-      episodes.set(routine.automationId, { ...routine, taskId })
-    },
-    deleteRoutineEpisode: async (automationId: string) => {
-      episodes.delete(automationId)
-    },
-  }
-  return {
-    created,
-    prompts,
-    delivered,
-    episodes,
-    inbox,
-    deps: {
-      store: args.store,
-      inbox,
-      orch: {
-        createTask:
-          args.createTask ??
-          (async (input: unknown) => {
-            created.push(input)
-            return { id: "task-1" } as DaemonTask
-          }),
-        getTask: (id: string) => args.tasks?.[id],
-      },
-      runtime: {
-        startTaskSessionWithPrompt:
-          args.start ??
-          (async (_l: DaemonRpcClient, _id: string, prompt: string) => {
-            prompts.push(prompt)
-            return { started: true }
-          }),
-        deliverPromptToLiveEngineDetailed:
-          args.deliver ??
-          (async (_task: unknown, prompt: string) => {
-            delivered.push(prompt)
-            return { outcome: "delivered" as const, tabId: "tab-1" }
-          }),
-      },
-      link,
-      now: () => NOW,
-    },
-  }
-}
+import { HOUR, NOW, REPO, automation, fakeDeps, tempStore } from "./automation-runner-fixtures.ts"
 
 describe("dueAutomations", () => {
   it("selects only enabled schedules whose time has arrived", () => {
@@ -400,116 +304,5 @@ describe("startAutomationRunner", () => {
     const { deps } = fakeDeps({ store })
     const stop = startAutomationRunner(deps, 60_000)
     expect(() => stop()).not.toThrow()
-  })
-})
-
-/**
- * A schedule is the only thing in Rove that acts with nobody watching, so the
- * queue that says "this needs you" is the only place its failures can land.
- * These pin the two rules that make that queue readable rather than noise.
- */
-describe("routine failures in the Inbox", () => {
-  it("files one episode per ROUTINE, however many firings fail", async () => {
-    const store = await tempStore()
-    const { deps, episodes } = fakeDeps({
-      store,
-      start: async () => ({ started: false, error: "engine process never started" }),
-    })
-    const a = automation()
-    for (let firing = 0; firing < 10; firing += 1) {
-      await runAutomationOnce(deps as never, a, { scheduledFor: NOW, trigger: "scheduled" })
-    }
-    // A fresh-task routine mints a new task every firing, so an episode keyed
-    // on the task would be 10 rows here — and 1,440 a day for a broken
-    // per-minute schedule.
-    expect(episodes.size).toBe(1)
-    expect(episodes.get(a.id)).toMatchObject({ name: "audit", status: "dispatch_failed" })
-  })
-
-  it("clears the episode once the routine runs cleanly again", async () => {
-    const store = await tempStore()
-    const { deps, episodes } = fakeDeps({
-      store,
-      start: async () => ({ started: false, error: "engine process never started" }),
-    })
-    const a = automation()
-    await runAutomationOnce(deps as never, a, { scheduledFor: NOW, trigger: "scheduled" })
-    expect(episodes.size).toBe(1)
-
-    const healthy = fakeDeps({ store, start: async () => ({ started: true }) })
-    // Same episode map: a fixed routine must not keep a permanent scar.
-    healthy.episodes.set(a.id, { name: "audit", status: "dispatch_failed", taskId: null })
-    await runAutomationOnce(healthy.deps as never, a, { scheduledFor: NOW, trigger: "scheduled" })
-    expect(healthy.episodes.size).toBe(0)
-  })
-
-  it("stays quiet for a precheck that found nothing to do", async () => {
-    const store = await tempStore()
-    const { deps, episodes } = fakeDeps({ store })
-    // `false` exits non-zero: healthy, and filing it would train the user to
-    // ignore the queue.
-    await runAutomationOnce(deps as never, automation({ precheck: { command: "false", timeoutSeconds: 5 } }), {
-      scheduledFor: NOW,
-      trigger: "scheduled",
-    })
-    expect(episodes.size).toBe(0)
-  })
-})
-
-/**
- * The sweep is serial and its ticker drops re-entrant ticks, so one slow
- * precheck stalls every routine behind it. The stall is survivable; a history
- * that does not mention it is not — `latestCronAtOrBefore` returns only the
- * newest occurrence, so the ones passed over used to leave no trace at all.
- */
-describe("occurrences the sweep never reached", () => {
-  it("records how many were passed over, and when the first was due", async () => {
-    const store = await tempStore()
-    // Armed for 09:55; the sweep only gets here at 10:00, so 09:55 (the one it
-    // was armed for) through 09:59 are occurrences that will never run — five
-    // of a per-minute schedule. 10:00 is the one it does run.
-    const created = await store.create({
-      name: "minute",
-      repo: REPO,
-      prompt: "p",
-      schedule: "* * * * *",
-      missedRunGraceMinutes: 60,
-    })
-    await store.update(created.id, {})
-    writeFileSync(
-      (store as unknown as { path: string }).path,
-      JSON.stringify({
-        version: 1,
-        automations: [{ ...created, nextRunAt: new Date(NOW - 5 * 60_000).toISOString() }],
-        runs: [],
-      }),
-    )
-    await store.init()
-
-    const { deps } = fakeDeps({ store })
-    await sweepAutomations(deps as never)
-
-    const runs = store.runsFor(created.id)
-    const gap = runs.find((run) => run.status === "skipped_missed")
-    expect(gap?.error).toBe(`5 earlier occurrences never ran (from ${new Date(NOW - 5 * 60_000).toISOString()})`)
-    // The occurrence it DID reach still runs — the gap row is a record, not a
-    // replacement for the firing.
-    expect(runs.some((run) => run.status === "dispatched")).toBe(true)
-  })
-
-  it("stays silent when the sweep arrived on time", async () => {
-    // Armed one minute ago, swept now: the healthy case, and the one that must
-    // not grow a second row.
-    const store = await tempStore(() => NOW - 60_000)
-    const created = await store.create({
-      name: "minute",
-      repo: REPO,
-      prompt: "p",
-      schedule: "* * * * *",
-      missedRunGraceMinutes: 60,
-    })
-    const { deps } = fakeDeps({ store })
-    await sweepAutomations(deps as never)
-    expect(store.runsFor(created.id).map((run) => run.status)).toEqual(["dispatched"])
   })
 })
